feat(wxapi): 升级公共组件

+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/CgibinExpress/Intracity/CgibinExpressIntracityQueryOrderResponse.cs
@@ -66,7 +66,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                 [Newtonsoft.Json.JsonProperty("lng")]
                 [System.Text.Json.Serialization.JsonPropertyName("lng")]
                 [System.Text.Json.Serialization.JsonNumberHandling(System.Text.Json.Serialization.JsonNumberHandling.AllowReadingFromString)]
-                public double Longitude { get; set; }
+                public decimal Longitude { get; set; }
 
                 /// <summary>
                 /// 获取或设置收件地址纬度。
@@ -74,7 +74,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                 [Newtonsoft.Json.JsonProperty("lat")]
                 [System.Text.Json.Serialization.JsonPropertyName("lat")]
                 [System.Text.Json.Serialization.JsonNumberHandling(System.Text.Json.Serialization.JsonNumberHandling.AllowReadingFromString)]
-                public double Latitude { get; set; }
+                public decimal Latitude { get; set; }
 
                 /// <summary>
                 /// 获取或设置收件人联系电话。
@@ -113,7 +113,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                 [Newtonsoft.Json.JsonProperty("lng")]
                 [System.Text.Json.Serialization.JsonPropertyName("lng")]
                 [System.Text.Json.Serialization.JsonNumberHandling(System.Text.Json.Serialization.JsonNumberHandling.AllowReadingFromString)]
-                public double Longitude { get; set; }
+                public decimal Longitude { get; set; }
 
                 /// <summary>
                 /// 获取或设置门店纬度。
@@ -121,7 +121,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Api.Models
                 [Newtonsoft.Json.JsonProperty("lat")]
                 [System.Text.Json.Serialization.JsonPropertyName("lat")]
                 [System.Text.Json.Serialization.JsonNumberHandling(System.Text.Json.Serialization.JsonNumberHandling.AllowReadingFromString)]
-                public double Latitude { get; set; }
+                public decimal Latitude { get; set; }
 
                 /// <summary>
                 /// 获取或设置门店联系电话。
